refactor: remove unneeded files
