+++ src/components/index.ts
@@ -1,6 +1,5 @@
 export * from "./App/App";
-export * from "./Input/Input";
 export * from "./FormBody/FormBody";
-export * from "./Banner/Banner";
-export * from "./StatusChip/StatusChip";
+export * from "./Input/Input";
 export * from "./Progress/Progress";
+export * from "./StatusChip/StatusChip";
