+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * from "./counter.type";
 export * from "./project.type";
-export * from "./variant.type";
