--- src/types/variant.type.ts
@@ -1,3 +1,0 @@
-import { Palette } from "@mui/material";
-
-export type Variant = keyof Palette;
