+++ src/views/index.ts
@@ -1,4 +1,3 @@
 export * from "./add-project/AddProject";
-export * from "./login/Login";
-export * from "./projects/Projects";
 export * from "./project/Project";
+export * from "./projects/Projects";
