--- src/views/login/components/index.tsx
@@ -1,1 +1,0 @@
-export * from "./LoginForm/LoginForm";
